legacy-data-converter/requirements.txt
# Legacy Data Converter - Requirements

# RabbitMQ
pika==1.3.2

# MinIO
minio==7.2.0

# Text extraction
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
lxml==4.9.3

# NLP (optional but recommended)
spacy==3.7.2
transformers==4.36.0
torch==2.1.2

# Utilities
python-dateutil==2.8.2

# Development
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
